infra/pyproject.toml
[tool.poetry]
name = "bluemoxon-infra"
version = "0.1.0"
description = "AWS CDK infrastructure for BlueMoxon"
authors = ["Mark"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.170.0"
constructs = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^24.0.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
